+++ src/repository/PackTypeRepository.java
@@ -43,7 +43,7 @@ public class PackTypeRepository {
       return result;
   }
 
-  public PackType getPackTypeById(Integer packTypeId) {
+  public PackType getPackTypeById(Long packTypeId) {
     TypedQuery<PackType> query;
     query = em.createQuery("SELECT b from PackType b " +
             "where b.id=:id", PackType.class);
